Update image_encoder.py
Fix for Torchscript issue for saving the model.

Looking at the provided code snippet, it's clear that pad_hw is only defined when self.window_size > 0. However, if self.window_size is not greater than 0, the pad_hw variable does not get a definition, leading to an error if trying to save SAM with Torchscript.